Refactor RelationalShapedQueryCompilingExpressionVisitor
